complete a signout page

+++ src/redux/actions/authActions.js
@@ -65,9 +65,22 @@ export const isUserLoggedIn = () => {
 
 export const signOut = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstant.LOGOUT_REQUEST
-        })
+        dispatch({ type: authConstant.LOGOUT_REQUEST })
+        const res = await axiosInstance.post(`/admin/signout`)
+
+        if (res.status === 200) {
+            localStorage.clear();
+            dispatch({
+                type: authConstant.LOGOUT_SUCCESS
+            })
+        } else {
+            if (res.status === 400) {
+                dispatch({
+                    type: authConstant.LOGOUT_FAILURE,
+                    payload: { error: res.data.error }
+                })
+            }
+        }
+
     }
 }
